move queryFeaturesInPoint to maplibre file
